+++ src/Java/Activation/Swish.java
@@ -3,12 +3,12 @@ package Java.Activation;
 public class Swish implements IActivationFunction {
 
     @Override
-    public double output(double x) {
-        return x * (1 / (1 + Math.exp(-x)));
+    public float output(float x) {
+        return (float) (x * (1 / (1 + Math.exp(-x))));
     }
 
     @Override
-    public double outputDerivative(double x) {
-        return ((1 + Math.exp(-x)) + x * Math.exp(-x)) / Math.pow(1 + Math.exp(-x), 2);
+    public float outputDerivative(float x) {
+        return (float) (((1 + Math.exp(-x)) + x * Math.exp(-x)) / Math.pow(1 + Math.exp(-x), 2));
     }
 }
