+++ src/Java/Activation/Tanh.java
@@ -3,12 +3,12 @@ package Java.Activation;
 public class Tanh implements IActivationFunction {
 
     @Override
-    public double output(double x) {
-        return Math.tanh(x);
+    public float output(float x) {
+        return (float) Math.tanh(x);
     }
 
     @Override
-    public double outputDerivative(double x) {
-        return 1 - Math.pow(Math.tanh(x), 2);
+    public float outputDerivative(float x) {
+        return (float) (1 - Math.pow(Math.tanh(x), 2));
     }
 }
