+++ src/Java/Math/RandomGenerator.java
@@ -2,8 +2,8 @@ package Java.Math;
 
 public class RandomGenerator {
 
-    public static double randomDouble(double min, double max) {
-        return min + (max - min) * Math.random();
+    public static float randomFloat(float min, float max) {
+        return (float) (min + (max - min) * Math.random());
     }
 
     public static int randomInt(int min, int max) {
